refactor: formatter를 utils, Movies를 data 패키지로 분리
