+++ app/src/main/java/woowacourse/movie/utils/PriceFormatter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.utils
 
 import java.text.DecimalFormat
